refactor: extract point constructions to point/constructions.ts
Moves all 43 point construction classes (+ circumcircle pair they
depend on) from the 2036-line Constructions.ts to a new
point/constructions.ts (772 lines).

Constructions.ts is now 1364 lines. It imports point constructions
for use in the registration array — no re-export needed since no
external code imports individual construction classes.

CircumcircleConstruction and CircumcircleConstruction2d live in the
point file (not circle/) because CircumcenterConstruction extends
CircumcircleConstruction directly.

Build passes, 98 tests pass, coverage stable.
